Major refactoring for "source files"
Moved to ISource for generic source file handling. Most of the
core bundles should be working. srcmvn.com looks to be rebuilding
correctly with these changes.

Bonus side effect: A *lot* less work is being done on each pass.
Also, if changes are made to sources that use a data provider
they now just get to use that data for free without requiring
the data provider to reload all content it provides.

 * Closes #16
 * Closes #14