Fixes duplicated content while watching source
Setting content to the formatted content was causing issues on
future calls to format the same content. In that case the formatted
content was being formatted again... in some cases this just looked
weird and in some cases this caused a hard error.

Downside to this fix is that templates extending from another
template (almost always the case) will not actually see the updates
from the templates they extend.

That will be a bug for a different day.

Fixes #21