View directory should be optional
The magic here is now gone. The magic was bad and made things more complicated
than it needed to be. FlexibleExtensionFilesystemLoader no longer needs to know
about source directory at all. This is as it should have been to begin with.

Setting a new (or adding additional) view directories now requires you to use
`%sculpin.source_dir%` in the path manually.

This addresses some concerns from @beryllium about the fact that the
documentation implied that one does not need to have a _views directory.

Closes #77