Fixed a regression for source views not being excluded
A long standing issue has sort of finally come to a head. There is this weird
distinction between things that are "source directory relative" vs. "anywhere on
the filesystem." This is an issue when ignoring/excluding files.

So for example, the exclude/ignore filters are based on directory paths relative
to the source directory. In the case of `source/_views`, we want a full path for
Twig to be able to locate the templates. However, we also want the source filter
to ignore it.

The hack before c6072dbd8282a46938e16d6162e109c43bf18797 was to "automagically"
add the sculpin.source_dir value to the front of the path if the naked view
directory did not exist on disk. Making this less magic was a good thing, but it
did not take into account that this would force the source filter to now try to
match against a full path (looking for `source/_views` in paths that look more
like `_views`).

The fix is to separate these concepts. We now have `source_view_paths` and
`view_paths`. The `source_view_paths` are always treated as being rooted in the
source directory. The `view_paths` can be located anywhere. Anything stored in
`view_paths` is going to be considered completely static; there is no real
chance for them to be excluded or ignored from the source set since they are
expected to live outside.

Fixes #79