Changes to configuration files will regenerate site
This is an attempt to fix a long standing issue where changes to config data
(specifically config/sculpin_site*.yml configuration) does not get reflected
until sculpin regenerates itself.

This solution is somewhat simple in that it makes a few assumptions about how
dot-access-configuration works, particularly with placeholder resolver's cache.
It is possible that this may result in unexpected behaviour.

This was tested out by changing `assets_version` key in sculpin_site.yml and it
would cause the entire site to be regenerated and all pages had the new
`assets_version` value as expected.

This was also tested by creating a new file in the config directory and adding a
new file was picked up.

Removing a file is not currently handled so that behaviour is currently unknown.

This might not work well with imports. :( This definitely needs some work but it
should help a bit for now.

closes #87