Sorting proxy source collections
This allows for a user to be able to configure how a proxy source collection for
a content type is sorted. This requires adding services to
`app/config/sculpin_services.yml` but it should leave for a lot of flexibility.
We can start providing additional sorters so that people do not have to write
new code themselves.

Example:

To configure a content type to be sorted based on its title rather than the
default order, one would do the following:

    services:
        sculpin_content_types.types.posts.collection.sorter:
            class: Sculpin\Contrib\ProxySourceCollection\Sorter\MetaSorter
            arguments:
                - title

To sort in ascending instead of descending order:

    services:
        sculpin_content_types.types.posts.collection.sorter:
            class: Sculpin\Contrib\ProxySourceCollection\Sorter\MetaSorter
            arguments:
                - title
                - desc