Add Nix flake for local development
This commit adds a Nix flake to use [Nix and nixpkgs][0] to create a
reproducible development environment for Sculpin, and not rely on
specifically installed versions by individual contributors or within CI
pipelines.

With these files in place, running `nix develop` will create a
development shell with the packages defined in flake.nix - locked to the
versions defined in flake.lock.

The currently included packages are:

* PHP 8.4.15
* Composer 2.9.2

There are [multiple versions of PHP][1] available in nixpkgs, if another
version would be better - but the PHPUnit tests and PHPStan checks all
pass within this shell.

[0]: https://nixos.org
[1]: https://search.nixos.org/packages?channel=unstable&query=php